+++ services/src/main/java/org/keycloak/broker/provider/HardcodedAttributeMapper.java
@@ -17,11 +17,13 @@
 
 package org.keycloak.broker.provider;
 
+import org.keycloak.models.Constants;
 import org.keycloak.models.IdentityProviderMapperModel;
 import org.keycloak.models.IdentityProviderSyncMode;
 import org.keycloak.models.KeycloakSession;
 import org.keycloak.models.RealmModel;
 import org.keycloak.models.UserModel;
+import org.keycloak.protocol.ProtocolMapperUtils;
 import org.keycloak.provider.ProviderConfigProperty;
 
 import java.util.ArrayList;
@@ -52,7 +54,9 @@ public class HardcodedAttributeMapper extends AbstractIdentityProviderMapper {
         property.setName(ATTRIBUTE_VALUE);
         property.setLabel("User Attribute Value");
         property.setHelpText("Value you want to hardcode");
-        property.setType(ProviderConfigProperty.STRING_TYPE);
+        property.setType(ProviderConfigProperty.MULTIVALUED_STRING_TYPE);
+        property.setStringify(Boolean.TRUE);
+        property.setDefaultValue("");
         configProperties.add(property);
     }
 
@@ -94,15 +98,16 @@ public class HardcodedAttributeMapper extends AbstractIdentityProviderMapper {
     @Override
     public void preprocessFederatedIdentity(KeycloakSession session, RealmModel realm, IdentityProviderMapperModel mapperModel, BrokeredIdentityContext context) {
         String attribute = mapperModel.getConfig().get(ATTRIBUTE);
-        String attributeValue = mapperModel.getConfig().get(ATTRIBUTE_VALUE);
-        context.setUserAttribute(attribute, attributeValue);
+        List<String> attributeValues = new ArrayList<>();
+        attributeValues.addAll(Arrays.asList(mapperModel.getConfig().get(ATTRIBUTE_VALUE).split(Constants.CFG_DELIMITER)));
+        context.setUserAttribute(attribute, attributeValues);
     }
 
     @Override
     public void updateBrokeredUser(KeycloakSession session, RealmModel realm, UserModel user, IdentityProviderMapperModel mapperModel, BrokeredIdentityContext context) {
         String attribute = mapperModel.getConfig().get(ATTRIBUTE);
-        String attributeValue = mapperModel.getConfig().get(ATTRIBUTE_VALUE);
-        user.setSingleAttribute(attribute, attributeValue);
+        List<String> attributeValues = Arrays.asList(mapperModel.getConfig().get(ATTRIBUTE_VALUE).split(Constants.CFG_DELIMITER));
+        user.setAttribute(attribute, attributeValues);
     }
 
     @Override